Extract diagnostics recorder ownership
